pyproject.toml
[project]
name = "bookstore-db"
version = "0.1.0"
description = ""
authors = [
    {name = "wydfaxil7",email = "a.faazil07@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "groq (>=1.1.2,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (<5.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]
